refactor(spanner): dedicated stub factory
