+++ google/cloud/spanner/google_cloud_cpp_spanner.bzl
@@ -88,6 +88,7 @@ google_cloud_cpp_spanner_hdrs = [
     "internal/session_pool.h",
     "internal/spanner_auth.h",
     "internal/spanner_stub.h",
+    "internal/spanner_stub_factory.h",
     "internal/status_utils.h",
     "internal/transaction_impl.h",
     "internal/tuple_utils.h",
@@ -174,6 +175,7 @@ google_cloud_cpp_spanner_srcs = [
     "internal/session_pool.cc",
     "internal/spanner_auth.cc",
     "internal/spanner_stub.cc",
+    "internal/spanner_stub_factory.cc",
     "internal/status_utils.cc",
     "internal/transaction_impl.cc",
     "keys.cc",
