+++ google/cloud/spanner/spanner_client_unit_tests.bzl
@@ -46,7 +46,7 @@ spanner_client_unit_tests = [
     "internal/route_to_leader_test.cc",
     "internal/session_pool_test.cc",
     "internal/spanner_auth_test.cc",
-    "internal/spanner_stub_test.cc",
+    "internal/spanner_stub_factory_test.cc",
     "internal/status_utils_test.cc",
     "internal/transaction_impl_test.cc",
     "internal/tuple_utils_test.cc",
